enable direct use of the waftool
Adding the tools directory in the script itself saves specifying
tooldir='waftools' in each call to ctx.load. Prepending
PYTHONPATH=waftools: to each call to waf on the command line would also
be possible but make the build process look more complicated if favor of
maintaining a wscript that poses as an exact example of how to use the
tool in other projects. This way, two lines are added at its top that
are easy to remove and do not hurt much when forgotten.